terminada nueva versión de la prueba de Velocidad con frecuencia de muestreo, frecuencia espacial y velocidad en la ventana de configuración. Falta har lo mismo con las otras pruebas que heredan de enrejado

+++ src/quino/clases/config/ConfigEnsayoEnrejado.java
@@ -19,20 +19,20 @@ public class ConfigEnsayoEnrejado extends ConfigEnsayoSinusoide {
         super();
     }
 
-    public ConfigEnsayoEnrejado(int direccion, int ppi, boolean onMove) {
+    public ConfigEnsayoEnrejado(int direccion, double ppi, boolean onMove) {
         super(direccion, ppi);
         this.onMove = onMove;
         this.key = onMove ? 32 : 0;
     }
 
-    public ConfigEnsayoEnrejado(int direccion, int ppi, boolean onMove, double contrat, double intensidadMedia) {
+    public ConfigEnsayoEnrejado(int direccion, double ppi, boolean onMove, double contrat, double intensidadMedia) {
         super(direccion, ppi, contrat, intensidadMedia);
         this.onMove = onMove;
         this.key = onMove ? 32 : 0;
     }
 
     public ConfigEnsayoEnrejado(int key, int panelEstimulo, int fs, double fspa_cpi_x, 
-            double fspa_cpi_y, int ppi, double fspa_cpp_x, double fspa_cpp_y, 
+            double fspa_cpi_y, double ppi, double fspa_cpp_x, double fspa_cpp_y, 
             boolean sentidoUpLeft, int direccion, boolean onMove, double contrat,
             double intensidadMedia) {
         super(key, panelEstimulo, fs, fspa_cpi_x, fspa_cpi_y, ppi, fspa_cpp_x,
